Card rewritten to TS
