+++ src/components/Card/Card.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import Card from '@components/Card'
+import { Card } from '@components/Card'
 import ColorNumber from '@components/ColorNumber'
 
 # Card
