+++ src/components/Card/Card.test.tsx
@@ -1,18 +1,16 @@
 import '@testing-library/jest-dom'
 
-import Card from '@components/Card'
+import { Card } from '@components/Card'
 import { Paragraph } from '@components/Paragraph'
 import { render } from '@testing-library/react'
-import React from 'react'
 
 describe('<Card /> mount', () => {
-  const props = {
-    variant: 'alert',
-    children: <Paragraph>Alert text</Paragraph>
-  }
-
   it('should display proper text', () => {
-    const { getByText } = render(<Card {...props} />)
+    const { getByText } = render(
+      <Card variant='alert'>
+        <Paragraph>Alert text</Paragraph>
+      </Card>
+    )
 
     const paragraph = getByText('Alert text')
 
@@ -20,7 +18,11 @@ describe('<Card /> mount', () => {
   })
 
   it('should display alert variant', () => {
-    const { getByTestId } = render(<Card {...props} />)
+    const { getByTestId } = render(
+      <Card variant='alert'>
+        <Paragraph>Alert text</Paragraph>
+      </Card>
+    )
 
     const card = getByTestId('card')
