+++ src/components/index.ts
@@ -5,11 +5,11 @@ export * from './Back'
 export * from './Backdrop'
 export * from './Badge'
 export { default as BlockSection } from './BlockSection'
-export { default as BoxBackground } from './BoxBackground'
+export * from './BoxBackground'
 export * from './BoxOutline'
 export { default as Button } from './Button'
 export * from './ButtonGroup'
-export { default as Card } from './Card'
+export * from './Card'
 export { default as Check } from './Check'
 export { default as CheckboxNoFormik } from './Checkbox'
 export * from './Close'
